Add setting to only find generated features
This might help with performance of structure and feature lookups
(e.g. with explorer maps or ender eyes) when not pre-generating a
map but might at the same time lead to structures not being found
by the items more often than normal when used in less-generated area.
Thanks @electronicboy for the improvements.