Add ability to find features with paper's async chunks
This adds a World#findNearestMapFeatureAsync method and a config option
to use the async search when creating treasure maps instead of blocking
the main thread with loading and generating chunks.

This has the slight downside that the items might break if interacted
with before the async search is finished. That should normally not occur,
especially on pre-generated worlds.