Fix wrong Nether Portal PlayerTeleportEvent
This fixes an issue where the player position is changed multiple times
while using a nether portal as well as calling the PlayerTeleportEvent
multiple times, even once with the wrong world and with the UNKNOWN cause.

Also fixes an issue with the target portal rotation not getting applied
on teleporting.