Use seed to check for features (Fixes PaperMC/Paper#2312)
This is done by using the same seed checking functionality that is used
by the server when generating these features instead of loading all
chunks around the location to check if they have the feature. The chunk
is only loaded after the generator confirmed a feature should be there.

The only downside of this is that it breaks once the seed or generator changes
but this should usually not happen but a config option is available anyways.